Kernel/riscv64: Move virtual memory definitions out of PageDirectory.h
This allows us to use them in the EFIPrekernel.
